+++ luna_workspace/condense_engine/src/helper_classes/DatabasePerst.java
@@ -111,7 +111,7 @@ public class DatabasePerst extends Database {
 	 * store a vector by values
 	 * 
 	 */
-	public void store(int data, int row, int col, double time) {
+	public void store(int data, int row, int col, int time) {
 		//TODO vectors.add( new GriddedVector( data, row, col, time));
 	}
